Pull qc.utils into pf.utils
* src/pf/utils.lua: Add choose and choose_index.

* tests/pfquickcheck/pflua_ir.lua:
* tests/pfquickcheck/prop_opt_eq_unopt.lua:
* tests/pfquickcheck/pflang_math.lua:
* tests/pfquickcheck/pflua-quickcheck.lua: Update to use choose from
  pf.utils.

* tests/pfquickcheck/qc/utils.lua: Remove.
